requirements.txt
pandas==1.1.5 ## 1.3.5
upsetplot
matplotlib==3.1.2
bokeh
numpy==1.18.1
scikit-learn==0.22.1
scipy==1.4.1
umap-learn==0.3.10
#reval it is present in the folder (with the specific old version used for this project)